requirements.txt
# Dependencias principales para LangChain y Azure OpenAI
langchain>=0.2.0
langchain-openai>=0.1.8
langgraph>=0.1.0

# LangGraph CLI y herramientas de desarrollo
langgraph-cli>=0.1.0

# Servidor para desarrollo (usado por langgraph dev)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Manejo de variables de entorno
python-dotenv>=1.0.0

# Validación de datos
pydantic>=2.5.0

# Dependencias adicionales para Azure OpenAI
openai>=1.12.0

# Utilidades para desarrollo y testing
typing-extensions>=4.8.0

# Google API para Google Tasks
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.88.0

# Herramientas adicionales para búsqueda web
tavily-python>=0.3.0

# Twilio para WhatsApp Business
twilio>=8.2.0

# LangGraph SDK para el servidor
langgraph-sdk>=0.1.0

# Utilidades adicionales
requests>=2.28.0
